VIRO-1510 seekToTime should take float vs int (renderer)
For both android and iOS

+++ ViroRenderer/VROVideoSurface.cpp
@@ -46,7 +46,7 @@ void VROVideoSurface::pause() {
     _texture->pause();
 }
 
-void VROVideoSurface::seekToTime(int seconds) {
+void VROVideoSurface::seekToTime(float seconds) {
     _texture->seekToTime(seconds);
 }
